PSEUDO FORCE TRIM added
it functions much like it is done in simulators, but with an interesting mode added to it. With an ADAPTIVE mode on, the stick will freeze when you press the trim button, and when you press it next time, it will only move again after your stick will be within a given range from its trimmed position! So now you can actually use it in flight in DCS.