Avoid uncomment Serial.print statements to calibrate sensors
Currently, to calibrate a sensor, a comment in the code
says that the user has to uncomment several Serial.print
statements. Also, it's not said in the code comments,
but the user also has to uncomment the Serial.begin
statement.

Thanks to C/C++ there's a easier way of doing this: with
the help of the preprocessor. Just by commenting or
uncommenting a #define, the preprocessor can insert or
not several lines of code.

This is the usual way of doing this kind of thing with
C/C++.

I have just inserted a #define CALIBRATE or
The user has only to comment/uncomment this simple
line of code.