Replace hard coded I2C address numbers by names
The I2C address are hardcoded as a number in every module of
master_controller. If one day the user wants to change the
I2C address of some device, it has to search the code for
the appropiate lines to edit. This can be complicated for
non-experienced users and can lead to errors.

This commit replaces the hardcoded I2C address numbers,
defining preprocessor macros in a_configuration.ino (in
master_controller sketch) and in the .ino files of every
individual device.

Users only have to edit the included #define's to change
the I2C addresses to the desired values.