Implements is_device_connected() to avoid repeating code
All the master_controller devices share a common code in
their setup function to check if the I2C device is connected.
To avoid repeating this code again and again on each device
the function is_device_connected() is defined in
z_shared_functions.ino. Every setup function calls this
function to check if an I2C device is connected.