radio panel knob pushbuttons are now mapped to 3 different sets of buttons depending on stack selector switch position
this is changed to add the possibility of mapping the audio panel mute / unmute buttons for com and nav. You can assign 2 buttons to ACT / STBY buttons for the corresponding radios, and use 2 other buttons for mute / unmute buttons on the radio panel. The 3rd set is there just in case.