Save RAM identifying connected devices.
In b_main.ino there are 11 boolean variables ("dev_*") created to store if
a devices are connected. Several bytes can be saved if a bit array is
used instead.

I have replaced the "dev_*" variables with a single 16 bits unsigned int
named "connected_devices". Each bit says if a device is connected (1) or
not (0). To identify the devices a new enum is created with entries in
the form "DEVICE_*".